requirements.txt
opencv-python-headless
mediapipe
numpy
matplotlib
python-dotenv
requests
sqlite3
